Fill small boring disconnected areas in the builder
This is effective in many cases, but while vaults default to
non-transparent, it is risky: the builder relies on an opaque vault's
exit being connected to something that is not masked as opaque, and uses
that something as a proxy for level connectivity in general. So if an
opaque vault is connected to a small space that this change would fill
in, that vault gets disconnected from the rest of the world. Until
vaults default to transparent, this commit therefore shouldn't be
merged.